style: split tensor_like to mod folder
